refactor(layer provider): Refactor
- Refactored hooks to their file
